refactor(viewer): split mod.rs into focused submodules
The monolithic src/viewer/mod.rs grew to 2610 lines mixing CLI parsing,
server state, parquet metadata extraction, HTTP routing, and a dozen
action handlers. Split it along those seams:

- mod.rs (~620): entry, command(), Config, run(), serve(), tests
- state.rs (~310): AppState, LazySectionStore, ProxyState, ApiResponse,
  CaptureParam, promql_error_type
- metadata.rs (~420): parquet metadata, multi-node info, file checksum,
  service-extension lookup, validate_service_extensions,
  regenerate_dashboards
- routes.rs (~430): app() router + read-side handlers (about, data,
  mode, sections, file_metadata, systeminfo, selection, query,
  query_range, labels, metadata, static assets)
- actions.rs (~630): mutating handlers (upload, load_url, attach/detach
  experiment, connect_agent, reset_tsdb, save_parquet,
  save_with_selection) + ingest_loop

Code-reuse improvements folded into the split:
- run_query helper consolidates capture lookup + QueryEngine for the
  two PromQL handlers
- snapshots_to_parquet centralizes live-mode parquet conversion shared
  by save_parquet and save_with_selection
- fetch_agent_info shared between CLI live-mode startup and the
  runtime /api/v1/connect handler
- parquet_attachment / server_error helpers cut response-builder
  boilerplate
- ApiResponse::err / ApiResponse::ok return pre-wrapped Json,
  removing ~30 verbose call sites
- run() decomposed into init_file_mode, init_live_mode,
  validate_category_at_startup, attach_cli_experiment, log_service_exts

Tests pass unchanged. No behavior changes.
